sharesense/requirements.txt
flask==3.1.0
pyjwt==2.10.1
flask-cors==5.0.1
